lots of changes to the menu class and a new csv file with fewer entries and fixed encoding

+++ PokemonApp/src/main/java/com/pokemonapp/pokemon/Pokemon.java
@@ -38,15 +38,15 @@ public class Pokemon {
         return description;
     }
     
-    public float GetHp() {
+    public int GetHp() {
         return hp;
     }
 
-    public float GetAttack() {
+    public int GetAttack() {
         return attack;
     }
 
-    public float Getdefense() {
+    public int Getdefense() {
         return defense;
     }
